extract options out of irust
